+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add an 'auto width' toggle for category labels that, when enabled, adjusts the label width to display all categories without trimming
 * Add settings to control axis display format and precision
 * Add settings to control bar size and to show percent of completion
+* Add "Legend" settings card
 
 ### Code improvements
 * Fix sync axis by computing min and max category values